Unify settings identifier sources
