+++ src/shared/lib/llm-provider/index.ts
@@ -1,5 +1,7 @@
 export { BaseLlmProvider } from './base-llm-provider'
-export type { LlmProviderId, ModelPurpose } from './base-llm-provider'
+export { LLM_PROVIDER_IDS } from './provider-types'
+export type { LlmProviderId } from './provider-types'
+export type { ModelPurpose } from './base-llm-provider'
 export { AnthropicLlmProvider } from './anthropic-provider'
 export { OpenRouterLlmProvider } from './openrouter-provider'
 export { BedrockLlmProvider } from './bedrock-provider'
@@ -30,7 +32,8 @@ export {
   resolveModelForProvider,
 } from './model-catalog'
 
-import type { LlmProviderId, ModelPurpose } from './base-llm-provider'
+import type { LlmProviderId } from './provider-types'
+import type { ModelPurpose } from './base-llm-provider'
 import { BaseLlmProvider } from './base-llm-provider'
 import type { ModelDefinition } from './model-catalog-schema'
 import { getEffectiveCatalog, getProviderCatalog, resolveModelForProvider } from './model-catalog'
