+++ src/shared/lib/llm-provider/provider-types.ts
@@ -1,0 +1,9 @@
+export const LLM_PROVIDER_IDS = [
+  'anthropic',
+  'openrouter',
+  'bedrock',
+  'platform',
+  'generic',
+] as const
+
+export type LlmProviderId = (typeof LLM_PROVIDER_IDS)[number]
